+++ src/main_state/mod.rs
@@ -1,2 +1,5 @@
 
-pub mod mouse;
+pub mod io;
+pub mod ms;
+pub mod loop_draw;
+pub mod loop_update;
