--- src/main_state/mouse.rs
@@ -1,1 +1,0 @@
-pub mod mouse;
